+++ ops/conv/index.ts
@@ -3,7 +3,11 @@ export {
   conv1dOutputLength,
   conv2d,
   conv2dOutputSize,
+  conv3d,
+  conv3dOutputSize,
   type Conv1dArgs,
   type Conv2dArgs,
   type Conv2dSpatial,
+  type Conv3dArgs,
+  type Conv3dSpatial,
 } from "./reference.js";